tools/requirements.txt
# Shared Python dependencies for all tools
beautifulsoup4>=4.12.0
python-slugify>=8.0.0
pandas>=2.1.0
pyyaml>=6.0.1

# Audit-specific dependencies
requests>=2.31.0
readability-lxml>=0.8.1
rapidfuzz>=3.5.0
lxml>=4.9.0
urllib3>=2.0.0

# Ingest-specific dependencies
markdownify>=0.11.6
python-frontmatter>=1.0.0
